inventory-service/requirements.txt
fastapi
uvicorn
kafka-python
databases[sqlite]
confluent-kafka
psycopg2-binary